Rework public software component API to handle mapping to Maven
When an adhoc component is created, we need a way to handle how
it should be published to Maven repositories. This commit reworks
the API so that when an adhoc variant is added, the plugin author
can tell to what Maven scope this should be published. Gradle
takes care of ordering the variants properly so that the generated
file is valid.